requirements.txt
torch~=2.0.1
torchvision~=0.15.2
tqdm~=4.65.0
webdataset~=0.2.48
scipy~=1.10.1
numpy
Pillow
timm
matplotlib